+++ docs/src/dev-docs/changelog.rst
@@ -24,6 +24,13 @@ changelog <https://keepachangelog.com/en/1.1.0/>`_ format. This project follows
 Unreleased
 ----------
 
+Changed
+#######
+
+- Metatrain is now built on top of ``metatomic`` instead of
+  ``metatensor.torch.atomistic``. Please refer to https://docs.metatensor.org/metatomic/
+  to find how to use the new models.
+
 Version 2025.6 - 2025-04-28
 ---------------------------